Move SSR code that is only used in test to test.
