+++ src/archive/gov/noaa/pfel/coastwatch/util/PipeToOutputStream.java
